feat(kotoba-wasm): browser-signed IPNS head record (commitHeadSigned) + wasm-safe kotoba-ipns-record
Implements the browser side of the content-addressed feed head (etzhayyim/root
ADR-2606066000): the browser node emits the SAME canonical, member-signed
IpnsRecord that kotoba-ipfs verifies — no parallel head format.

- New crate `kotoba-ipns-record`: the IpnsRecord/IpnsName/IpnsRegistryError types
  + Ed25519 sign/verify, factored out of kotoba-ipfs so they are wasm32-clean
  (deps: serde, ciborium, ed25519-dalek, multibase, ipld-core, thiserror — no
  tokio/reqwest). Adds `IpnsRecord::with_value_string` so a caller that already
  holds the head CID as a multibase string (the browser) need not pull the CID
  crate. 5 unit tests (sign→verify roundtrip, tamper-value, tamper-sequence,
  unsigned, browser-shaped-record-verifies). Builds for BOTH host and
  wasm32-unknown-unknown.
- kotoba-ipfs: re-exports the moved types (`pub use kotoba_ipns_record::…`); the
  native registries (InMemory/Persistent/Kubo/Signed) are unchanged. All 12 ipns
  tests still green (registries sign/verify through the re-exported type =
  byte-identical).
- kotoba-wasm: `WriteCrypto::sign_ipns_head(name, value_multibase, sequence,
  valid_until)` builds + signs a canonical IpnsRecord with the member key, and a
  new `#[wasm_bindgen] commitHeadSigned(sequence, valid_until)` that commits +
  returns the signed record JSON. Supersedes the `{root,did,sig}` shim / trusted
  KV `kroot:` value: the record is self-verifying, so the apex relays it
  untrusted (no-server-key).

Verified: kotoba-ipns-record host+wasm32 build + 5/5 tests; kotoba-ipfs host
build + 12/12 ipns tests; kotoba-wasm host build clean. The full kotoba-wasm
wasm32 LINK is blocked by a PRE-EXISTING tokio/mio transitive dep via
kotoba-kqe/kotoba-datomic (unrelated to this change; the wasm-pack bundle path
handles it) — the crate added here is verified wasm32-clean in isolation.
