+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
     "crates/kotoba-signal",
     "crates/kotoba-ingest",
     "crates/kotoba-ipfs",
+    "crates/kotoba-ipns-record",
     "crates/kotoba-cli",
     "crates/kotoba-edn",
     "crates/kotoba-datomic",
