+++ crates/kotoba-ipfs/Cargo.toml
@@ -18,6 +18,9 @@ tracing.workspace   = true
 ed25519-dalek.workspace = true
 multibase.workspace = true
 
+# Wasm-safe IPNS record type (shared with kotoba-wasm; ADR-2606066000).
+kotoba-ipns-record = { path = "../kotoba-ipns-record" }
+
 # IPFS-compatible CID substrate plus a lightweight self-owned block exchange.
 ipld-core = { workspace = true }
 multihash-codetable = { version = "0.2", features = ["sha2"] }
