+++ crates/kotoba-ipns-record/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "kotoba-ipns-record"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+description = "Wasm-safe kotoba IPNS head record (member-signed, content-addressed-head pointer) — shared by kotoba-ipfs (native registries) and kotoba-wasm (browser publish). ADR-2606066000."
+
+# Pure-Rust, wasm32-clean: NO tokio/reqwest. The native IPNS registries
+# (InMemory/Persistent/Kubo) stay in kotoba-ipfs; only the record TYPE + its
+# Ed25519 sign/verify live here so the browser node can emit a canonical,
+# self-verifying head record (no parallel format — ADR-2605262130).
+[dependencies]
+serde = { workspace = true }
+ciborium = { workspace = true }
+ed25519-dalek = { workspace = true }
+multibase = { workspace = true }
+ipld-core = { workspace = true }
+thiserror = { workspace = true }
