+++ crates/kotoba-wasm/Cargo.toml
@@ -16,6 +16,9 @@ kotoba-core = { path = "../kotoba-core" }
 kotoba-datomic = { path = "../kotoba-datomic" }
 kotoba-edn = { path = "../kotoba-edn" }
 kotoba-kqe = { path = "../kotoba-kqe" }
+# Wasm-safe canonical IPNS head record (ADR-2606066000) — the browser emits the
+# SAME member-signed record kotoba-ipfs verifies (no parallel format).
+kotoba-ipns-record = { path = "../kotoba-ipns-record" }
 serde = { workspace = true }
 serde_json = { workspace = true }
 ciborium = { workspace = true }
